Delete command trait for taxonomy and metabox command
